convert simple functions into ES6 function
