+++ src/index.js
@@ -1,10 +1,7 @@
+/* eslint-disable no-unused-vars */
 import './style.css';
 import getdata from './module/getdata.js';
-
-const gameName = document.getElementById('gameName');
-const gameScore = document.getElementById('gameScore');
-const submit = document.getElementById('btn');
-
+import postdata from './module/postdata.js';
 
 const getGameId = async () => {
   const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games';
@@ -25,24 +22,3 @@ const getGameId = async () => {
 if (!localStorage.getItem('gameId')) {
   getGameId();
 }
-
-async function postData(gameId) {
-  const url = `https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/${gameId}/scores`;
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      user: gameName.value,
-      score: gameScore.value,
-    }),
-  });
-  return response.json();
-}
-
-const getGameIdFromLocalStorage = localStorage.getItem('gameId');
-getdata(getGameIdFromLocalStorage);
-submit.addEventListener('click', () => {
-  postData(getGameIdFromLocalStorage);
-});
