+++ src/module/getdata.js
@@ -1,5 +1,7 @@
-export default async function getdata(gameId) {
+const getdata = async (gameId) => {
   const response = await fetch(`https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/${gameId}/scores`);
-  const responseFromApi = await response.json();
-  console.log('getdata', responseFromApi);
-}
+  // const data = await response.json();
+  // console.log(data);
+  return response.json();
+};
+export default getdata;
